ImageIO fixes - possible overflow errors with Uint32 - switched to size_t
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@873 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ java/nitf/src/java/nitf/BlockingInfo.java
@@ -104,14 +104,14 @@ public final class BlockingInfo extends DestructibleObject
      * 
      * @return
      */
-    public native int getLength();
+    public native long getLength();
 
     /**
      * Sets the length
      * 
      * @param length
      */
-    public native void setLength(int length);
+    public native void setLength(long length);
 
     /**
      * Prints the data pertaining to this object to an OutputStream
